Work on CreateVacancy scene, created new componenent DatePicker
